refactor: restructure webix view into jet view for popup
