+++ sources/views/activities.js
@@ -2,6 +2,7 @@ import {JetView} from "webix-jet";
 import {activities} from "../models/activities";
 import {activityTypes} from "../models/activityTypes";
 import {contacts} from "../models/contacts";
+import PopupView from "./popupView";
 
 export default class ActivitiesView extends JetView {
 	config() {
@@ -14,47 +15,6 @@ export default class ActivitiesView extends JetView {
 			{id: "6", value: "This week"},
 			{id: "7", value: "This month"}
 		];
-
-		const popupForm = {
-			view: "form",
-			id: "popup_form",
-			elements: [
-				{view: "text", label: "Details"},
-				{
-					view: "select",
-					label: "Type",
-					options: activityTypes
-				},
-				{
-					view: "select",
-					label: "Contact",
-					options: contacts
-				},
-				{cols: [
-					{view: "datepicker", label: "Date", value: new Date()},
-					{view: "datepicker", type: "time", label: "Time", value: new Date()}
-				]},
-				{view: "checkbox", labelRight: "Selected"},
-				{cols: [
-					{gravity: 4},
-					{view: "button", value: "Add/delete", css: "webix_primary"},
-					{view: "button", value: "Cancel"}
-				]}
-			]
-		};
-
-		const popup = webix.ui({
-			view: "window",
-			localid: "popup",
-			move: true,
-			head: "Edit/Add activity",
-			width: 600,
-			height: 400,
-			position: "center",
-			close: true,
-			modal: true,
-			body: popupForm
-		});
 
 		return {
 			rows: [
@@ -72,8 +32,9 @@ export default class ActivitiesView extends JetView {
 						label: "Add activity",
 						type: "icon",
 						icon: "mdi mdi-plus-box",
+						css: "webix_primary",
 						click: () => {
-							popup.show();
+							this.ui(PopupView).showWindow();
 						}
 					}
 				]},
@@ -121,5 +82,4 @@ export default class ActivitiesView extends JetView {
 }
 
 // columns options
-// select 0th option from select view
 // position of checkbox with labelRight property
